Rename NyPlanSide components folder
