+++ src/components/NyPlanSide/OppsummeringSteg/NarDuFerdigstillerPlanenAlert.tsx
@@ -15,9 +15,9 @@ export default function NarDuFerdigstillerPlanenAlert({ className }: Props) {
         Når du velger å ferdigstille planen deles den med den ansatte, og de kan
         finne den ved å logge seg på nav.no. Du som arbeidsgiver vil også finne
         planen under Oppfølgingsplaner på nav.no. Hvis du ikke er klar til å
-        dele den med den ansatte enda, kan du lagre planen som utkast, og du
-        finner den igjen under oppfølgingsplaner for den ansatte. Den er da
-        markert som utkast.
+        dele planen med den ansatte enda, så kan du avslutte og fortsette
+        senere. Det du har skrevet finner du igjen i et utkast under
+        oppfølgingsplaner for den ansatte.
       </BodyLong>
     </Alert>
   );
