+++ src/components/NyPlanSide/OppsummeringSteg/OppsummeringSteg.tsx
